refactor: gamblers 패키지 이동
